requirements.txt
keyboard
psutil
PyQt5
pystray
Pillow